requirements_dev.txt
-e .
pip==8.1.2
bumpversion==0.5.3
wheel==0.29.0
watchdog==0.8.3
flake8==2.6.0
tox==2.3.1
coverage==4.1
Sphinx==1.4.8
coreapi-cli==1.0.6
cumulusci==2.0.0b68
nose>=1.3.7
nose-tap>=1.9
requests==2.18.4
